# Use 0 for empty value