--- a/simd/match.s
+++ b/simd/match.s
@@ -4,11 +4,11 @@
 
 #include "textflag.h"
 
-// func MatchMetadata(metadata *[16]int8, hash int8) uint16
+// func MatchMetadata(metadata *[16]uint8, hash uint8) uint16
 // Requires: SSE2, SSSE3
 TEXT ·MatchMetadata(SB), NOSPLIT, $0-18
 	MOVQ     metadata+0(FP), AX
-	MOVBLSX  hash+8(FP), CX
+	MOVBLZX  hash+8(FP), CX
 	MOVD     CX, X0
 	PXOR     X1, X1
 	PSHUFB   X1, X0