--- a/simd/match_amd64.go
+++ b/simd/match_amd64.go
@@ -6,4 +6,4 @@ package simd
 
 // MatchMetadata performs a 16-way probe of |metadata| using SSE instructions
 // nb: |metadata| must be an aligned pointer
-func MatchMetadata(metadata *[16]int8, hash int8) uint16
+func MatchMetadata(metadata *[16]uint8, hash uint8) uint16