--- a/simd/match_amd64.go
+++ b/simd/match_amd64.go
@@ -7,3 +7,7 @@ package simd
 // MatchMetadata performs a 16-way probe of |metadata| using SSE instructions
 // nb: |metadata| must be an aligned pointer
 func MatchMetadata(metadata *[16]uint8, hash uint8) uint16
+
+// MatchMetadata performs a 16-way probe of zero byte using SSE instructions
+// nb: |metadata| must be an aligned pointer
+func MatchEmpty(metadata *[16]uint8) uint16