--- a/simd/src/asm.go
+++ b/simd/src/asm.go
@@ -23,22 +23,41 @@ import (
 func main() {
 	ConstraintExpr("amd64")
 
-	TEXT("MatchMetadata", NOSPLIT, "func(metadata *[16]uint8, hash uint8) uint16")
-	Doc("MatchMetadata performs a 16-way probe of |metadata| using SSE instructions",
-		"nb: |metadata| must be an aligned pointer")
-	m := Mem{Base: Load(Param("metadata"), GP64())}
-	h := Load(Param("hash"), GP32())
-	mask := GP32()
+	{
+		TEXT("MatchMetadata", NOSPLIT, "func(metadata *[16]uint8, hash uint8) uint16")
+		Doc("MatchMetadata performs a 16-way probe of |metadata| using SSE instructions",
+			"nb: |metadata| must be an aligned pointer")
+		m := Mem{Base: Load(Param("metadata"), GP64())}
+		h := Load(Param("hash"), GP32())
+		mask := GP32()
 
-	x0, x1, x2 := XMM(), XMM(), XMM()
-	MOVD(h, x0)
-	PXOR(x1, x1)
-	PSHUFB(x1, x0)
-	MOVOU(m, x2)
-	PCMPEQB(x2, x0)
-	PMOVMSKB(x0, mask)
+		x0, x1, x2 := XMM(), XMM(), XMM()
+		MOVD(h, x0)
+		PXOR(x1, x1)
+		PSHUFB(x1, x0)
+		MOVOU(m, x2)
+		PCMPEQB(x2, x0)
+		PMOVMSKB(x0, mask)
 
-	Store(mask.As16(), ReturnIndex(0))
-	RET()
+		Store(mask.As16(), ReturnIndex(0))
+		RET()
+	}
+
+	{
+		TEXT("MatchEmpty", NOSPLIT, "func(metadata *[16]uint8) uint16")
+		Doc("MatchMetadata performs a 16-way probe of zero byte using SSE instructions",
+			"nb: |metadata| must be an aligned pointer")
+		m := Mem{Base: Load(Param("metadata"), GP64())}
+		mask := GP32()
+
+		x0, x1 := XMM(), XMM()
+		PXOR(x0, x0)
+		MOVOU(m, x1)
+		PCMPEQB(x1, x0)
+		PMOVMSKB(x0, mask)
+
+		Store(mask.As16(), ReturnIndex(0))
+		RET()
+	}
 	Generate()
 }