--- a/simd/src/go.mod
+++ b/simd/src/go.mod
@@ -1,0 +1,10 @@
+module github.com/dolthub/swiss/simd/src
+
+go 1.22.4
+
+require github.com/mmcloughlin/avo v0.6.0
+
+require (
+	golang.org/x/mod v0.14.0 // indirect
+	golang.org/x/tools v0.16.1 // indirect
+)