… simd/src/go.sum