In the current implementation we have to initialize all metadata values to the specific `empty` value when we're creating a new map or growing it.

This changes empty to be the zero byte, and tombstone to be byte 1, by adding 2 to all hashes.

We do have to run one extra sum `+2` to ensure that `h2` hashes don't have the lower bit set, but we compensate that by doing one less operation when matching `empty` in the group, (two less assembly instructions in the amd64 simd code!), so there's no penalty there. 

I benchmarked on ARM (Apple Silicon), and for some reason some other benchmark results have also improved (we're doing some less work, but not that much less. Maybe some operations are optimized? but why int benchmarks are unaffected then?), I'll run more benchmarks to confirm, but it would be great if someone on amd64 could double check.

```
goos: darwin
goarch: arm64
pkg: github.com/dolthub/swiss
                                   │    main-20    │             empty-zero-20             │
                                   │    sec/op     │    sec/op      vs base                │
MatchMetadata-12                     0.4646n ±  0%   0.4652n ±  0%        ~ (p=0.291 n=20)
StringMaps/n=16/runtime_map-12        6.098n ±  1%    6.172n ±  2%        ~ (p=0.092 n=20)
StringMaps/n=16/swiss.Map-12          13.95n ±  4%    12.67n ±  3%   -9.21% (p=0.000 n=20)
StringMaps/n=128/runtime_map-12       6.345n ±  2%    6.365n ±  1%        ~ (p=0.597 n=20)
StringMaps/n=128/swiss.Map-12         14.87n ±  3%    13.99n ±  2%   -5.95% (p=0.000 n=20)
StringMaps/n=1024/runtime_map-12      6.830n ±  1%    6.818n ±  1%        ~ (p=0.995 n=20)
StringMaps/n=1024/swiss.Map-12        11.22n ±  8%    11.93n ±  6%   +6.33% (p=0.027 n=20)
StringMaps/n=8192/runtime_map-12      13.02n ±  8%    13.22n ±  7%        ~ (p=0.941 n=20)
StringMaps/n=8192/swiss.Map-12        14.30n ± 15%    12.25n ±  7%  -14.40% (p=0.010 n=20)
StringMaps/n=131072/runtime_map-12    20.45n ±  5%    20.44n ±  0%        ~ (p=0.663 n=20)
StringMaps/n=131072/swiss.Map-12      15.77n ±  2%    15.49n ±  2%        ~ (p=0.351 n=20)
Int64Maps/n=16/runtime_map-12         4.471n ±  1%    4.530n ±  2%        ~ (p=0.159 n=20)
Int64Maps/n=16/swiss.Map-12           4.651n ± 22%    4.756n ± 26%   +2.28% (p=0.003 n=20)
Int64Maps/n=128/runtime_map-12        4.614n ±  1%    4.621n ±  2%        ~ (p=0.673 n=20)
Int64Maps/n=128/swiss.Map-12          5.423n ±  2%    5.481n ±  3%        ~ (p=0.277 n=20)
Int64Maps/n=1024/runtime_map-12       4.733n ±  1%    4.737n ±  1%        ~ (p=0.995 n=20)
Int64Maps/n=1024/swiss.Map-12         5.663n ±  1%    5.845n ±  1%   +3.21% (p=0.001 n=20)
Int64Maps/n=8192/runtime_map-12       6.762n ± 10%    7.054n ±  6%        ~ (p=0.331 n=20)
Int64Maps/n=8192/swiss.Map-12         6.556n ±  1%    6.691n ±  1%   +2.07% (p=0.000 n=20)
Int64Maps/n=131072/runtime_map-12     17.12n ±  0%    16.74n ±  0%   -2.22% (p=0.000 n=20)
Int64Maps/n=131072/swiss.Map-12      10.155n ±  1%    9.874n ±  2%   -2.77% (p=0.000 n=20)
NewMap/n=16-12                        96.66n ±  0%    90.18n ±  0%   -6.70% (p=0.000 n=20)
NewMap/n=128-12                       315.1n ±  0%    270.5n ±  0%  -14.14% (p=0.000 n=20)
NewMap/n=1024-12                      1.616µ ±  0%    1.128µ ±  1%  -30.23% (p=0.000 n=20)
NewMap/n=8192-12                     11.731µ ±  1%    5.293µ ±  0%  -54.88% (p=0.000 n=20)
NewMap/n=131072-12                   159.87µ ±  1%    48.96µ ±  1%  -69.37% (p=0.000 n=20)
Map_Put_Growing/n=16-12               214.2n ±  0%    203.5n ±  0%   -5.00% (p=0.000 n=20)
Map_Put_Growing/n=128-12              1.399µ ±  0%    1.319µ ±  0%   -5.72% (p=0.000 n=20)
Map_Put_Growing/n=1024-12             11.08µ ±  0%    10.45µ ±  0%   -5.65% (p=0.000 n=20)
Map_Put_Growing/n=8192-12             90.19µ ±  0%    84.92µ ±  0%   -5.85% (p=0.000 n=20)
Map_Put_Growing/n=131072-12           1.646m ±  2%    1.559m ±  2%   -5.26% (p=0.000 n=20)
geomean                               63.68n          57.82n         -9.20%
```