@colega 2024-07-04
I ran the benchmarks in a Github codespace, the results make (more) sense:
```
goos: linux
goarch: amd64
pkg: github.com/dolthub/swiss
cpu: AMD EPYC 7763 64-Core Processor                
                                  │ /home/codespace/main.bench │   /home/codespace/empty-zero.bench   │
                                  │           sec/op           │    sec/op     vs base                │
MatchMetadata-2                                   2.881n ±  6%   2.807n ±  4%        ~ (p=0.063 n=10)
StringMaps/n=16/runtime_map-2                     10.48n ±  9%   10.24n ±  6%        ~ (p=0.670 n=10)
StringMaps/n=16/swiss.Map-2                       16.88n ±  7%   17.45n ±  5%        ~ (p=0.342 n=10)
StringMaps/n=128/runtime_map-2                    10.92n ±  6%   10.80n ±  8%        ~ (p=0.954 n=10)
StringMaps/n=128/swiss.Map-2                      18.09n ±  5%   17.83n ±  3%        ~ (p=0.579 n=10)
StringMaps/n=1024/runtime_map-2                   11.43n ± 16%   11.21n ±  9%        ~ (p=0.579 n=10)
StringMaps/n=1024/swiss.Map-2                     18.68n ±  5%   19.25n ±  8%        ~ (p=0.197 n=10)
StringMaps/n=8192/runtime_map-2                   25.03n ±  3%   24.78n ±  5%        ~ (p=0.644 n=10)
StringMaps/n=8192/swiss.Map-2                     22.12n ±  6%   22.66n ±  5%        ~ (p=0.436 n=10)
StringMaps/n=131072/runtime_map-2                 33.67n ±  3%   34.12n ±  4%        ~ (p=1.000 n=10)
StringMaps/n=131072/swiss.Map-2                   36.48n ±  7%   37.34n ±  5%        ~ (p=0.912 n=10)
Int64Maps/n=16/runtime_map-2                      7.904n ±  6%   7.636n ±  6%        ~ (p=0.353 n=10)
Int64Maps/n=16/swiss.Map-2                        14.46n ±  4%   14.43n ±  5%        ~ (p=0.971 n=10)
Int64Maps/n=128/runtime_map-2                     8.245n ±  6%   8.073n ±  5%        ~ (p=0.280 n=10)
Int64Maps/n=128/swiss.Map-2                       15.18n ±  5%   15.32n ±  2%        ~ (p=0.631 n=10)
Int64Maps/n=1024/runtime_map-2                    8.336n ±  7%   8.523n ±  9%        ~ (p=0.256 n=10)
Int64Maps/n=1024/swiss.Map-2                      15.49n ±  8%   16.02n ±  4%        ~ (p=0.382 n=10)
Int64Maps/n=8192/runtime_map-2                    20.08n ±  5%   18.47n ±  9%   -7.99% (p=0.029 n=10)
Int64Maps/n=8192/swiss.Map-2                      16.99n ±  4%   16.86n ±  7%        ~ (p=0.362 n=10)
Int64Maps/n=131072/runtime_map-2                  28.42n ±  2%   27.52n ±  3%   -3.17% (p=0.019 n=10)
Int64Maps/n=131072/swiss.Map-2                    30.38n ±  2%   30.07n ±  2%        ~ (p=0.172 n=10)
NewMap/n=16-2                                     248.2n ± 15%   219.2n ±  6%  -11.68% (p=0.000 n=10)
NewMap/n=128-2                                    779.5n ±  7%   684.0n ±  9%  -12.25% (p=0.001 n=10)
NewMap/n=1024-2                                   3.614µ ± 12%   2.743µ ±  7%  -24.10% (p=0.000 n=10)
NewMap/n=8192-2                                   50.49µ ± 49%   37.31µ ±  7%        ~ (p=0.089 n=10)
NewMap/n=131072-2                                 328.9µ ±  8%   240.6µ ± 10%  -26.83% (p=0.000 n=10)
Map_Put_Growing/n=16-2                            571.1n ±  6%   527.5n ± 12%   -7.64% (p=0.043 n=10)
Map_Put_Growing/n=128-2                           3.405µ ± 14%   3.389µ ± 12%        ~ (p=0.481 n=10)
Map_Put_Growing/n=1024-2                          26.59µ ±  7%   26.23µ ±  7%        ~ (p=0.247 n=10)
Map_Put_Growing/n=8192-2                          235.5µ ±  4%   236.1µ ±  4%        ~ (p=0.796 n=10)
Map_Put_Growing/n=131072-2                        4.814m ± 10%   4.484m ±  5%        ~ (p=0.579 n=10)
geomean                                           142.1n         135.8n         -4.47%
```